Move login / register to pages directory
